requirements.txt

numpy>=1.20.0,<2.3.0  
drake>=1.46.0
matplotlib>=3.5.0  

